refactor(itemlist,store): remove usless code, move sse parser into middleware
