--- src/modules/oh-indicators-list/List.tsx
@@ -1,66 +1,0 @@
-import React, { useEffect } from 'react';
-import { useSelector, useDispatch } from 'react-redux';
-import styled from 'styled-components';
-
-import { actions, selectors } from './store';
-
-const Root = styled.section`
-  width: 100vw;
-`;
-
-const UnorderedList = styled.ul`
-  list-style-type: none;
-  display: flex;
-  flex-wrap: wrap;
-  gap: 10px;
-`;
-
-const ListItem = styled.li`
-  height: 75px;
-  width: 150px;
-  font-size: 12px;
-  border-radius: 5px;
-  box-shadow: 2px 2px 15px 1px rgb(0 0 0 / 10%);
-  padding: 5px;
-  box-sizing: border-box;
-  overflow: hidden;
-`;
-
-const Headline = styled.h3`
-  font-weight: 700;
-  overflow: hidden;
-  text-overflow: ellipsis;
-`;
-
-const Value = styled.div``;
-
-export const List = () => {
-  const items = useSelector(selectors.selectItems);
-  const error = useSelector(selectors.selectErrorMessage);
-  const dispatch = useDispatch();
-
-  useEffect(() => {
-    dispatch(actions.getAllItems());
-    dispatch(actions.initSseConnection());
-    return () => {
-      dispatch(actions.closeSseConnection());
-    };
-  }, []);
-
-  return (
-    <Root className="App">
-      {error}
-      <UnorderedList>
-        {Object.keys(items).map((itemKey) => {
-          const { name, state } = items[itemKey];
-          return (
-            <ListItem key={name}>
-              <Headline>{name}</Headline>
-              <Value>{state}</Value>
-            </ListItem>
-          );
-        })}
-      </UnorderedList>
-    </Root>
-  );
-};
