+++ src/modules/oh-indicators-list/index.ts
@@ -1,3 +1,2 @@
-export { List } from './List';
 export * as indicatorsStore from './store';
 export { default as SseClient } from './network-layer/sseClient';
